+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.67.18]
+
+### Fixed
+
+- last set_context event is now sent within a last message
+
+## [3.67.17]
+
+### Fixed
+
+- added missing JSDOC
+
 ## [3.67.16]
 
 ### Fixed
@@ -46,7 +58,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - replaced vulnerable version of axios
-
 
 ## [3.67.9]